backend/requirements.txt
fastapi
uvicorn
pinecone-client
google-generativeai
gdown
python-dotenv
pydantic
httpx
